# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The found record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        },
        "summary": "Get a user by ID",
        "tags": [
          "Users"
        ]
      }
    }
  },
  "info": {
    "title": "User Management API",
    "description": "API documentation for the User Management service.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Users",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The full name of the user.",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "description": "The user's email address.",
            "example": "jane.doe@example.com"
          },
          "role": {
            "type": "string",
            "description": "The role of the user.",
            "enum": [
              "admin",
              "user"
            ],
            "default": "user"
          }
        },
        "required": [
          "name",
          "email",
          "role"
        ]
      },
      "User": {
        "type": "object",
        "properties": {}
      }
    }
  }
}